+++ src/main/java/kr/hs/entrydsm/husky/domain/application/domain/repositories/async/GeneralApplicationAsyncRepository.java
@@ -1,4 +1,4 @@
-package kr.hs.entrydsm.husky.domain.application.domain.repositories.generalapplication;
+package kr.hs.entrydsm.husky.domain.application.domain.repositories.async;
 
 import kr.hs.entrydsm.husky.domain.application.domain.adapter.GeneralApplicationAdapter;
