+++ src/main/java/kr/hs/entrydsm/husky/domain/application/domain/repositories/async/GeneralApplicationAsyncRepositoryImpl.java
@@ -1,7 +1,6 @@
-package kr.hs.entrydsm.husky.domain.application.domain.repositories.generalapplication;
+package kr.hs.entrydsm.husky.domain.application.domain.repositories.async;
 
 import kr.hs.entrydsm.husky.domain.application.domain.adapter.GeneralApplicationAdapter;
-import kr.hs.entrydsm.husky.domain.application.domain.repositories.async.ApplicationAsyncRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
